setup webpack, started pulling out scripts from the field templates
