+++ package.json
@@ -1,0 +1,16 @@
+{
+  "name": "CRUD",
+  "dependencies": {
+    "scriptjs": "^2.5.8"
+  },
+  "devDependencies": {
+    "babel-cli": "^6.14.0",
+    "babel-loader": "^6.2.5",
+    "babel-preset-es2015": "^6.14.0",
+    "webpack": "^1.13.2"
+  },
+  "scripts": {
+    "dev": "webpack -d --watch",
+    "prod": "webpack -p"
+  }
+}
