+++ src/resources/views/fields/select2.blade.php
@@ -40,23 +40,6 @@
         <link href="{{ asset('vendor/backpack/select2/select2-bootstrap-dick.css') }}" rel="stylesheet" type="text/css" />
     @endpush
 
-    {{-- FIELD JS - will be loaded in the after_scripts section --}}
-    @push('crud_fields_scripts')
-        <!-- include select2 js-->
-        <script src="{{ asset('vendor/backpack/select2/select2.js') }}"></script>
-        <script>
-            jQuery(document).ready(function($) {
-                // trigger select2 for each untriggered select2 box
-                $('.select2').each(function (i, obj) {
-                    if (!$(obj).data("select2"))
-                    {
-                        $(obj).select2();
-                    }
-                });
-            });
-        </script>
-    @endpush
-
 @endif
 {{-- End of Extra CSS and JS --}}
 {{-- ########################################## --}}
