+++ src/resources/views/fields/summernote.blade.php
@@ -24,17 +24,6 @@
         <link href="{{ asset('vendor/backpack/summernote/summernote.css') }}" rel="stylesheet" type="text/css" />
     @endpush
 
-    {{-- FIELD JS - will be loaded in the after_scripts section --}}
-    @push('crud_fields_scripts')
-        <!-- include summernote js-->
-        <script src="{{ asset('vendor/backpack/summernote/summernote.min.js') }}"></script>
-        <script>
-            jQuery(document).ready(function($) {
-                $('.summernote').summernote();
-            });
-        </script>
-    @endpush
-
 @endif
 {{-- End of Extra CSS and JS --}}
 {{-- ########################################## --}}
