+++ src/resources/views/fields/table.blade.php
@@ -84,76 +84,6 @@
         {{-- YOUR CSS HERE --}}
     @endpush
 
-    {{-- FIELD JS - will be loaded in the after_scripts section --}}
-    @push('crud_fields_scripts')
-        {{-- YOUR JS HERE --}}
-        <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/angular.js/1.5.8/angular.min.js"></script>
-        <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/jqueryui/1.12.1/jquery-ui.min.js"></script>
-        <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/angular-ui-sortable/0.14.3/sortable.min.js"></script>
-        <script>
-            angular.module('backpackTable', ['ui.sortable'], function($interpolateProvider){
-                $interpolateProvider.startSymbol('<%');
-                $interpolateProvider.endSymbol('%>');
-            })
-            .controller('tableController', function($scope){
-
-                $scope.sortableOptions = {
-                    handle: '.sort-handle'
-                };
-
-                $scope.addItem = function(){
-
-                    if( $scope.max > -1 ){
-                        if( $scope.items.length < $scope.max ){
-                            var item = {};
-                            $scope.items.push(item);
-                        } else {
-                            new PNotify({
-                                title: $scope.maxErrorTitle,
-                                text: $scope.maxErrorMessage,
-                                type: 'error'
-                            });
-                        }
-                    }
-                    else {
-                        var item = {};
-                        $scope.items.push(item);
-                    }
-                }
-
-                $scope.removeItem = function(item){
-                    var index = $scope.items.indexOf(item);
-                    $scope.items.splice(index, 1);
-                }
-
-                $scope.$watch('items', function(a, b){
-
-                    if( $scope.min > -1 ){
-                        while($scope.items.length < $scope.min){
-                            $scope.addItem();
-                        }
-                    }
-
-                    if( typeof $scope.items != 'undefined' && $scope.items.length ){
-
-                        if( typeof $scope.field != 'undefined'){
-                            if( typeof $scope.field == 'string' ){
-                                $scope.field = $($scope.field);
-                            }
-                            $scope.field.val( angular.toJson($scope.items) );
-                        }
-                    }
-                }, true);
-
-                if( $scope.min > -1 ){
-                    for(var i = 0; i < $scope.min; i++){
-                        $scope.addItem();
-                    }
-                }
-            });
-        </script>
-
-    @endpush
 @endif
 {{-- End of Extra CSS and JS --}}
 {{-- ########################################## --}}
