+++ src/resources/views/fields/upload.blade.php
@@ -25,28 +25,3 @@
         <p class="help-block">{!! $field['hint'] !!}</p>
     @endif
 </div>
-
-{{-- FIELD EXTRA JS --}}
-{{-- push things in the after_scripts section --}}
-
-    @push('crud_fields_scripts')
-        <!-- no scripts -->
-        <script>
-	        $("#{{ $field['name'] }}_file_clear_button").click(function(e) {
-	        	e.preventDefault();
-	        	$(this).parent().addClass('hidden');
-
-	        	var input = $("#{{ $field['name'] }}_file_input");
-	        	input.removeClass('hidden');
-	        	input.attr("value", "").replaceWith(input.clone(true));
-	        	// add a hidden input with the same name, so that the setXAttribute method is triggered
-	        	$("<input type='hidden' name='{{ $field['name'] }}' value=''>").insertAfter("#{{ $field['name'] }}_file_input");
-	        });
-
-	        $("#{{ $field['name'] }}_file_input").change(function() {
-	        	console.log($(this).val());
-	        	// remove the hidden input, so that the setXAttribute method is no longer triggered
-	        	$(this).next("input[type=hidden]").remove();
-	        });
-        </script>
-    @endpush
